Analyse

Analyse

# S3 method for mb_model
analyse(
  x,
  data,
  nchains = getOption("mb.nchains", 3L),
  niters = getOption("mb.niters", 1000L),
  nthin = getOption("mb.thin", NULL),
  parallel = getOption("mb.parallel", FALSE),
  quiet = getOption("mb.quiet", TRUE),
  glance = getOption("mb.glance", TRUE),
  beep = getOption("mb.beep", TRUE),
  ...
)

Arguments

x
An object inheriting from class mb_model or a list of such objects.

data
The data frame to analyse.

nchains
A count of the number of chains.

niters
A count of the number of simulations to save per chain.

nthin
A count of the thining interval or NULL (in which case taken from model).

parallel
A flag indicating whether to perform the analysis in parallel if possible.

quiet
A flag indicating whether to disable tracing information.

glance
A flag indicating whether to print a model summary.

beep
A flag indicating whether to beep on completion of the analysis.

...
Additional arguments.