MB Model

Creates MB model.

# S3 method for mb_code
model(x, gen_inits = function(data) {     list() },
  random_effects = list(), fixed = getOption("mb.fixed", "^[^e]"),
  derived = character(0), select_data = list(), center = character(0),
  scale = character(0), modify_data = identity, niters = 10^3,
  new_expr = character(0), modify_new_data = identity, drops = list(),
  ...)

Arguments

x
An object inheriting from class mb_code.

gen_inits
A single argument function taking the modified data and returning a named list of initial values.

random_effects
A named list specifying the random effects and the associated factors.

fixed
A string of a regular expression specifying the fixed parameters to monitor.

derived
A character vector of the derived parameters to monitor.

select_data
A named list specifying the columns to select and their associated classes and values as well as transformations and scaling options.

center
A character vector of the columns to center.

scale
A character vector of the columns to scale (after centering).

modify_data
A single argument function to modify the data (in list form) immediately prior to the analysis.

niters
A number specifying the number of iterations to sample for each chain. Must be 10^3, 10^4, 10^5 or 10^6.

new_expr
A string of R code specifying the predictive relationships.

modify_new_data
A single argument function to modify new data (in list form) immediately prior to calculating new_expr.

drops
A list of character vector of possible scalar parameters to drop (fix at 0).

...
Unused arguments.

Value

An object inherting from class mb_model.

Details

For tmb models gen_inits must specify all the fixed parameters. Missing random parameters are assigned the value 0.

For jmb models unspecified the initial values for each chain are drawn from the prior distributions.

See also

check_data2 rescale_c